Move snapshot tests to a dedicated test target for better visibility
