+++ platforms/ios/example/Wysiwyg.xcodeproj/xcshareddata/xcschemes/Wysiwyg.xcscheme
@@ -53,6 +53,16 @@
             skipped = "NO">
             <BuildableReference
                BuildableIdentifier = "primary"
+               BlueprintIdentifier = "WysiwygComposerSnapshotTests"
+               BuildableName = "WysiwygComposerSnapshotTests"
+               BlueprintName = "WysiwygComposerSnapshotTests"
+               ReferencedContainer = "container:../lib/WysiwygComposer">
+            </BuildableReference>
+         </TestableReference>
+         <TestableReference
+            skipped = "NO">
+            <BuildableReference
+               BuildableIdentifier = "primary"
                BlueprintIdentifier = "A6472CC02886CF840021A0E8"
                BuildableName = "WysiwygUITests.xctest"
                BlueprintName = "WysiwygUITests"
